refactor(langfuse-tracing): bring server/observability/langfuse-tracing.ts to the lint standard (#780)
The file carried 8 oxlint findings: 982 code lines against 500, four
unicorn(no-useless-fallback-in-spread), and three complexity findings
(readMcpTracingConfig 13, maskTraceValue 17, buildToolTraceBody 22).

Split along the seams the file already had, each new module self-contained
and none of them a private copy of something that existed:

  trace-error-label.ts  content-free error labelling, shared by the wrapper
                        and the sink health tracker rather than copied
  trace-types.ts        the public type contracts, type-only so the SDK sink
                        and the wrapper can both import them without a cycle
  trace-release.ts      release identity; still defers the deploy-stamp parse
                        to readDeployedRevision in transport/server-identity.ts
  trace-masking.ts      credential masking over src/secret-patterns.ts, which
                        remains the owner of detection
  trace-sink-health.ts  the outage state machine and its two log lines
  trace-sink.ts         the only code touching @langfuse/* and @opentelemetry/*
  trace-body.ts         trace body construction, pure and SDK-free
  trace-config.ts       reading the env config and a call's session identity

langfuse-tracing.ts re-exports every moved symbol, so callers and the three
pinning test suites keep one import surface and stay unmodified.

The three complexity findings are cleared by decomposition, not by moving
behavior: maskTraceValue delegates arrays, built-in containers, and plain
objects to named helpers; buildToolTraceBody splits metadata, caller identity
(now a field table), and span masking; readMcpTracingConfig reads its three
coordinates through one trimmedEnv helper.

The four spread findings are cleared by collapsing traceRetrievalSpan and
traceRetrievalSpanSync, which were byte-identical apart from one await, onto a
shared recordRetrievalOutcome. Spreading undefined adds no keys, so dropping
the ?? {} is the same object. The ?? {} inside toolTraceMetadata is kept: it
feeds maskTraceValue as an argument, where it is not a spread and does matter.

errorOutput keeps its exported behavior exactly, since src/background-tracing.ts
imports it. No schema, default, error string, SQL, or log event name changed.

Verified: oxlint --deny-warnings over all eight files exits 0; bunx tsc --noEmit
clean; bun run test:isolated over langfuse-tracing.test.ts, config.test.ts, and
src/index.test.ts is 151 pass / 0 fail with those suites unmodified.
